Update README.md
I’m Pacific Byiringiro, founder of Uganda Eco Youths Network. This movement began in 2024 with my late friend Magana Praise, when we were still students. We saw the suffering in our communities especially among school children  and felt the need to act.

The journey has never been easy. We started with nothing but a vision. Many doubted us, and some still do. But I kept learning, connecting, and discovering one truth: even heroes start from zero.

Today, we are no longer just two  we are many. A rising team of passionate young people building sustainable change. This is just the beginning, and we welcome you to be part of it.